tvdt/pyproject.toml
[tool.poetry]
name = "tvdt"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.2"
django-crispy-forms = "^2.3"
crispy-bootstrap5 = "^2024.10"
django-allauth = {extras = ["socialaccount"], version = "^65.2.0"}
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.0"}
environs = {extras = ["django"], version = "^11.2.1"}
psycopg2 = "^2.9.10"

[tool.poetry.group.dev.dependencies]
mypy = "^1.11.0"
black = "^24.10.0"
isort = "^5.13.2"
djlint = "^1.36.3"

[tool.poetry.group.prod.dependencies]
gunicorn = "^23.0.0"

[tool.isort]
profile = "black"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "tvdt.settings"

[tool.djlint]
profile="django"
